nova-sonic/webapp/requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
loguru==0.7.3

# WebRTC and audio processing
pipecat-ai==0.0.67
pipecat-ai-small-webrtc-prebuilt==0.5.0
aiortc==1.11.0
av==14.4.0
pylibsrtp==0.12.0

# Utilities
aiohttp==3.11.18
httpx==0.28.1